[docs] Point temporary documentation to a single source of truth
Currently, every module in the Ethical Framework ecosystem is
listed on the README, but updating the README on every repo
whenever a dependency is added, removed, or renamed is too much
work for the author. This commits points the documentation to the
main repo for a list of all the other modules.